fix entity, add package and add lombok dependancy
