--- src/main/java/Address.java
@@ -1,7 +1,0 @@
-public class Address {
-
-    private int number;
-    private String address;
-    private String town;
-    private int zip;
-}
