--- src/main/java/Person.java
@@ -1,7 +1,0 @@
-public class Person {
-
-    private String name;
-    private int phone;
-    private int mobile;
-    private String email;
-}
