--- src/main/java/Print.java
@@ -1,5 +1,0 @@
-import java.util.List;
-
-public interface Print<T> {
-    public void print(List<T> p);
-}
