--- src/main/java/Remove.java
@@ -1,3 +1,0 @@
-public interface Remove {
-    public void remove();
-}
