--- src/main/java/Search.java
@@ -1,3 +1,0 @@
-public interface Search {
-    public void search();
-}
