+++ src/main/java/org/perscholas/data/Address.java
@@ -1,0 +1,16 @@
+package org.perscholas.data;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class Address {
+
+    private int number;
+    private String address;
+    private String town;
+    private int zip;
+}
