+++ src/main/java/org/perscholas/data/Print.java
@@ -1,0 +1,7 @@
+package org.perscholas.data;
+
+import java.util.List;
+
+public interface Print<T> {
+    void print(List<T> p);
+}
