+++ src/main/java/org/perscholas/data/Remove.java
@@ -1,0 +1,5 @@
+package org.perscholas.data;
+
+public interface Remove<T> {
+    void remove(T object);
+}
