+++ src/main/java/org/perscholas/data/Search.java
@@ -1,0 +1,5 @@
+package org.perscholas.data;
+
+public interface Search<T> {
+    void search(T object);
+}
